Solving Magento Bug: allowing numeric SKUs to be used
Still happening on Magento 1.9.2.

There's a bug in Magento API that doe not allow to get a Product by SKU when the SKU string can be transformed to a number.
There's a documented workaround on the following URL, basically consisting on adding a whitespace at the end that will be trimmed by Magento.

More info on:
http://stackoverflow.com/questions/6748142/magento-1-5-numeric-skus-and-productidentifiertype/10915276#10915276